sam-3d-lite/requirements.txt
torch>=2.0.0
transformers>=4.35.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
open3d>=0.17.0
matplotlib>=3.7.0
